--- src/test/java/codingbat/dastan/ParrotTrouble.java
@@ -1,13 +1,0 @@
-package codingbat.dastan;
-
-public class ParrotTrouble {
-    public boolean parrotTrouble(boolean talking, int hour) {
-        if(!talking)
-            return false;
-        else if(hour > 20 || hour < 7)
-            return true;
-        else
-            return false;
-
-    }
-}
