--- src/test/java/codingbat/dastan/SleepIn.java
@@ -1,13 +1,0 @@
-package codingbat.dastan;
-
-public class SleepIn {
-    public boolean sleepIn(boolean weekday, boolean vacation) {
-        if (!weekday || vacation) {
-            return true;
-        }
-
-        return false;
-    }
-
-
-}
