--- src/test/java/codingbat/dastan/SumDouble.java
@@ -1,11 +1,0 @@
-package codingbat.dastan;
-
-public class SumDouble {
-    public int sumDoubleMethod(int a, int b) {
-        if(a == b){
-            return (a + b) * 2;
-        }else{
-            return a + b;
-        }
-    }
-}
