--- src/test/java/codingbat/dastan/monkeyTrouble.java
@@ -1,14 +1,0 @@
-package codingbat.dastan;
-
-public class monkeyTrouble {
-    public boolean monkeyTroubleMethod(boolean aSmile, boolean bSmile) {
-        if(aSmile == true && bSmile == true){
-            return true;
-        }
-        else if (aSmile == false && bSmile == false){
-            return true;
-        }else{
-            return false;
-        }
-    }
-}
